Simple refactor to place all AllWinner boards within the same module
