+++ src/adafruit_blinka/microcontroller/allwinner/__init__.py
@@ -1,0 +1,1 @@
+"""Definition of all Allwinner chips"""
