+++ src/adafruit_blinka/microcontroller/allwinner/a64/__init__.py
@@ -1,0 +1,1 @@
+"""Definition for the AllWinner A64 chip"""
