+++ src/adafruit_blinka/microcontroller/allwinner/h3/__init__.py
@@ -1,0 +1,1 @@
+"""Definition for the AllWinner H3 chip"""
