+++ src/digitalio.py
@@ -16,7 +16,7 @@ if detector.chip.BCM2XXX:
 elif detector.chip.AM33XX:
     from adafruit_blinka.microcontroller.am335x.pin import Pin
 elif detector.chip.SUN8I:
-    from adafruit_blinka.microcontroller.allwinner_h3.pin import Pin
+    from adafruit_blinka.microcontroller.allwinner.h3.pin import Pin
 elif detector.chip.SAMA5:
     from adafruit_blinka.microcontroller.sama5.pin import Pin
 elif detector.chip.T210:
